backend/requirements.txt
# FastAPI Backend Requirements
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-settings==2.7.0

# ML/AI
sentence-transformers==5.2.0
faiss-cpu==1.13.2
torch==2.5.1
transformers==4.57.6
openai==2.15.0

# Image Processing
pillow==12.1.0
pytesseract==0.3.13

# Utilities
numpy==2.4.1
tqdm==4.67.1
requests==2.32.5
